added 2 new classes for integration testing, added new methods in search and graph, and updated comments

+++ src/project/Relocate/GraphEdge.java
@@ -8,15 +8,15 @@ package project.Relocate;
 public class GraphEdge implements Comparable{
 	private GraphVertex v;
 	private GraphVertex w;
-	private double weight;
+	private long weight;
 	
 	/**
 	 * Creates a new connection between two vertices
 	 * @param v The first vertex to connect
 	 * @param w The second vertex to connect
-	 * @param weight A value that stores the weight of the node
+	 * @param weight A long value that stores the weight of the node
 	 */
-	public GraphEdge(GraphVertex v, GraphVertex w, double weight){
+	public GraphEdge(GraphVertex v, GraphVertex w, long weight){
 		this.v = v;
 		this.w = w;
 		this.weight = weight;
@@ -39,9 +39,9 @@ public class GraphEdge implements Comparable{
 	
 	/**
 	 * Returns the weight of this edge
-	 * @return A value holding the weight of this edge
+	 * @return A long value holding the weight of this edge
 	 */
-	public double getWeight(){
+	public long getWeight(){
 		return weight;
 	}
